requirements.txt
# requirements.txt
# pip install -r requirements.txt
python-dotenv
qdrant-client
langchain
groq
langchain-text-splitters
pypdf
sentence-transformers